+++ Core/src/main/java/com/jasperls/rimor/data/ExecutionData.java
@@ -1,4 +1,4 @@
-package com.jasperls.rimor;
+package com.jasperls.rimor.data;
 
 import lombok.Getter;
 import lombok.Setter;
